app/requirements.txt
Flask
mysql-connector
opencv-python
pyaudio
wave
moviepy
